Switch to the one event listener for some buttons
Change JQuery listeners for buttons that should only be pressed once to "one" instead of "on". The one event listener only accepts one call and then unbinds the event listener. As such it means the function is only run once. Because the buttons I've changed to one call functions that change the view. Also changed the "Continue" button to use a JQuery "one" event listener instead of disabling the button for consistency.