If no archive notice is found don't error out
If no archive notice is found by spiHelperParseArchiveNotice the code ignores this and marches on to try and get the username from the result of the regex. If no match was found, null is returned so accessing the 2nd element of null will give an error. As such if the result is null there is no archive notice. The behavior I've written for after this to return the parsed archive notice type with everything as null to signify it doesn't exist (as opposed to empty) as some code relies on this format being returned. Using null for these cases still works as intended (as it doesn't match). Fixes #75 
Also add code to generate an archive notice to add to the page if it doesn't exist (along with __TOC__ and {{SPIpriorcases}} if they are not there). If it was removed by a vandal this would probably be reverted, so I don't think there is a case for not automatically adding it in. A warning will also be visually shown to say that it was added automatically.