Add a setting to not auto-tick the archive case
Add a setting to not auto-tick the archive case, which is set to by default tick the archive checkbox if the case is closed. If set to false the checkbox will not be automatically ticked. I would find this useful as when I archive I use the SPI-Archive link instead of the main SPI view. When I select a case I nearly always just want to comment on a closed case or re-open it.