Use checkuserblock-account when CU blocking
Use checkuserblock-account when CU blocking by default. Add a setting to override this and continue to use uw-sockblock when CU blocking.
{{checkuserblock-account}}, though less standard, does detail appealing to arbcom and also adds a explanation for sysops. I think this makes it better than {{uw-sockblock}} for CU blocks.
At the moment {{checkuserblock-account}} is not subst'ed, as this is how the template is usually used per the documentation. However, I have started a talk page discussion about possibly changing this. If that discussion has no opposition or support, I will BOLDly make the template substed and submit a separate patch.